classes.js create
class declaration add
